LPS-104952 Extracting logic to the canonical url provider
